feat: refactor region module and testing
