+++ .github/workflows/validate.yml
@@ -7,8 +7,13 @@ jobs:
   build:
     name: validate
     uses: aztfmods/module-azurerm-global/.github/workflows/validate.yml@main
+    with:
+      test: "node-pools"
+      workload: demo
+      environment: dev
     secrets:
       ARM_CLIENT_ID: ${{ secrets.ARM_CLIENT_ID }}
       ARM_CLIENT_SECRET: ${{ secrets.ARM_CLIENT_SECRET }}
       ARM_SUBSCRIPTION_ID: ${{ secrets.ARM_SUBSCRIPTION_ID }}
       ARM_TENANT_ID: ${{ secrets.ARM_TENANT_ID }}
+      AZURE_CREDENTIALS: ${{ secrets.AZURE_CREDENTIALS }}
