+++ examples/complete/main.tf
@@ -2,44 +2,50 @@ provider "azurerm" {
   features {}
 }
 
-module "global" {
-  source = "github.com/aztfmods/module-azurerm-global"
+module "region" {
+  source = "github.com/aztfmods/module-azurerm-regions"
 
-  company = "cn"
-  env     = "p"
-  region  = "weu"
+  workload    = var.workload
+  environment = var.environment
 
-  rgs = {
-    demo = { location = "westeurope" }
-  }
+  location = "westeurope"
+}
+
+module "rg" {
+  source = "github.com/aztfmods/module-azurerm-rg"
+
+  workload       = var.workload
+  environment    = var.environment
+  location_short = module.region.location_short
+  location       = module.region.location
 }
 
 module "analytics" {
   source = "github.com/aztfmods/module-azurerm-law"
 
-  company = module.global.company
-  env     = module.global.env
-  region  = module.global.region
+  workload       = var.workload
+  environment    = var.environment
+  location_short = module.region.location_short
 
   law = {
-    location      = module.global.groups.demo.location
-    resourcegroup = module.global.groups.demo.name
+    location      = module.rg.group.location
+    resourcegroup = module.rg.group.name
     sku           = "PerGB2018"
     retention     = 90
   }
-  depends_on = [module.global]
+  depends_on = [module.rg]
 }
 
 module "kv" {
   source = "github.com/aztfmods/module-azurerm-kv"
 
-  company = module.global.company
-  env     = module.global.env
-  region  = module.global.region
+  workload       = var.workload
+  environment    = var.environment
+  location_short = module.region.location_short
 
   vault = {
-    location      = module.global.groups.demo.location
-    resourcegroup = module.global.groups.demo.name
+    location      = module.rg.group.location
+    resourcegroup = module.rg.group.name
 
     secrets = {
       tls_public_key = {
@@ -56,20 +62,20 @@ module "kv" {
       }
     }
   }
-  depends_on = [module.global]
+  depends_on = [module.rg]
 }
 
 module "aks" {
   source = "../../"
 
-  company = module.global.company
-  env     = module.global.env
-  region  = module.global.region
+  workload       = var.workload
+  environment    = var.environment
+  location_short = module.region.location_short
 
   aks = {
-    location            = module.global.groups.demo.location
-    resourcegroup       = module.global.groups.demo.name
-    node_resource_group = "${module.global.groups.demo.name}-node"
+    location            = module.rg.group.location
+    resourcegroup       = module.rg.group.name
+    node_resource_group = "${module.rg.group.name}-node"
 
     enable = {
       public_access = true
@@ -123,5 +129,5 @@ module "aks" {
       }
     }
   }
-  depends_on = [module.global]
+  depends_on = [module.rg]
 }
