+++ examples/container-registry/main.tf
@@ -2,44 +2,50 @@ provider "azurerm" {
   features {}
 }
 
-module "global" {
-  source = "github.com/aztfmods/module-azurerm-global"
+module "region" {
+  source = "github.com/aztfmods/module-azurerm-regions"
 
-  company = "cn"
-  env     = "p"
-  region  = "weu"
+  workload    = var.workload
+  environment = var.environment
 
-  rgs = {
-    demo = { location = "westeurope" }
-  }
+  location = "westeurope"
+}
+
+module "rg" {
+  source = "github.com/aztfmods/module-azurerm-rg"
+
+  workload       = var.workload
+  environment    = var.environment
+  location_short = module.region.location_short
+  location       = module.region.location
 }
 
 module "acr" {
   source = "github.com/aztfmods/module-azurerm-acr"
 
-  company = module.global.company
-  env     = module.global.env
-  region  = module.global.region
+  workload       = var.workload
+  environment    = var.environment
+  location_short = module.region.location_short
 
   registry = {
-    location      = module.global.groups.demo.location
-    resourcegroup = module.global.groups.demo.name
+    location      = module.rg.group.location
+    resourcegroup = module.rg.group.name
     sku           = "Premium"
   }
-  depends_on = [module.global]
+  depends_on = [module.rg]
 }
 
 module "aks" {
   source = "../../"
 
-  company = module.global.company
-  env     = module.global.env
-  region  = module.global.region
+  workload       = var.workload
+  environment    = var.environment
+  location_short = module.region.location_short
 
   aks = {
-    location            = module.global.groups.demo.location
-    resourcegroup       = module.global.groups.demo.name
-    node_resource_group = "${module.global.groups.demo.name}-node"
+    location            = module.rg.group.location
+    resourcegroup       = module.rg.group.name
+    node_resource_group = "${module.rg.group.name}-node"
     dns_prefix          = "demo"
 
     registry = {
