+++ examples/simple/variables.tf
@@ -1,4 +1,2 @@
-variable "aks" {}
 variable "workload" {}
 variable "environment" {}
-variable "location_short" {}
