+++ tests/aks_test.go
@@ -1,70 +1,28 @@
 package main
 
 import (
-	"github.com/gruntwork-io/terratest/modules/terraform"
 	"os"
-	"path/filepath"
 	"testing"
+
+	"github.com/aztfmods/module-azurerm-aks/shared"
+	"github.com/gruntwork-io/terratest/modules/terraform"
 )
 
-var filesToCleanup = []string{
-	"*.terraform*",
-	"*tfstate*",
-}
-
-type TestCase struct {
-	name string
-	path string
-}
-
-func getTerraformOptions(terraformDir string) *terraform.Options {
-	return &terraform.Options{
-		TerraformDir: terraformDir,
-		NoColor:      true,
-		Parallelism:  20,
-	}
-}
-
-func cleanup(t *testing.T, tfOpts *terraform.Options) {
-	terraform.Destroy(t, tfOpts)
-	cleanupFiles(t, tfOpts.TerraformDir)
-}
-
-func TestApplyNoError(t *testing.T) {
+func TestAppyNoError() {
 	t.Parallel()
 
-	tests := []TestCase{
-		{name: "simple", path: "../examples/simple"},
-		{name: "node-pools", path: "../examples/node-pools"},
-		{name: "container-registry", path: "../examples/container-registry"},
-		{name: "complete", path: "../examples/complete"},
+	tests := []shared.TestCase{
+		{Name: os.Getenv("USECASE"), Path: "../examples/" + os.Getenv("USECASE")},
 	}
 
 	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			terraformOptions := getTerraformOptions(test.path)
+		t.Run(test.Name, func(t *testing.T) {
+			terraformOptions := shared.GetTerraformOptions(test.Path)
 
 			terraform.WithDefaultRetryableErrors(t, &terraform.Options{})
 
-			defer cleanup(t, terraformOptions)
+			defer shared.Cleanup(t, terraformOptions)
 			terraform.InitAndApply(t, terraformOptions)
 		})
-	}
-}
-
-func cleanupFiles(t *testing.T, dir string) {
-	for _, pattern := range filesToCleanup {
-		matches, err := filepath.Glob(filepath.Join(dir, pattern))
-		if err != nil {
-			t.Logf("Error: %v", err)
-			continue
-		}
-		for _, filePath := range matches {
-			if err := os.RemoveAll(filePath); err != nil {
-				t.Logf("Failed to remove %s: %v\n", filePath, err)
-			} else {
-				t.Logf("Successfully removed %s\n", filePath)
-			}
-		}
 	}
 }
